feat!: web socket server
The web socket server allows connections from the client for syncing.
Simply POSTing changes will also be an option. As part of this work
some missed details came to light that needed addressing.

In order to properly evaluate each incoming change, it's necessary to know
the value of the vclock at the time of the change. I knew the implementation
was necessary but somehow it got lost among all the other details. This
change adds that missing piece, as well as updating the `*_changes` and
`*_pending` tables to simplify some of the querying.