fix: changes incorrectly assessed as stale
TL;DR:

- ensure date objects are used in `VCompare.isOutDated()`
- update field used to set `*_record_meta.modified`
- removed stray `mod` column from `*_pending`
- broadcast changes on web socket connections
- improvement to the web socket implementation

Found that some consecutive changes would be rejected when syncing on
another client. The main culprit was in VCompare where it was actually
comparing a string to a date object (in some cases) — oops!

During the same investigation I realised that the record meta was
acquiring the modified time of the record meta's update, rather
than the actual change's modified time. There was also a straggling
`mod` column on the `*_pending` table that was causing an issue.
Between these various errors a little refactoring was required. The
C[R]UD triggers have been updated, along with the tests. Everything
is green.

All of the above was discovered during the course of implementing and
testing the web socket server with the client.